+++ lib/strava/models/athlete.rb
@@ -20,6 +20,14 @@ module Strava
       property 'follower'
       property 'friend'
       property 'summit'
+
+      def name
+        [firstname, lastname].compact.join(' ') if firstname || lastname
+      end
+
+      def strava_url
+        "https://www.strava.com/athletes/#{username || id}"
+      end
     end
   end
 end